Allows for setting map styles based on game data
With this you can set the background styles(and any other styles)
in the game data file. This allows for an image to be set for the
background, then you can set the transparency(using 0-1) and it
will set the opacity at that level so you can see the background
image.