add "match" color and globals for track color
This adds the ability to set track colors, width, etc. with globals in the
"info" section. Track can be made into a dashed line if the color and
borderColor are set the same as the hex color. However, it's inconvenient
to use them together since the hex color varies across the map and tiles.
This also adds a new color for track of "match" which means to match the
hex color.