adding "match" and global settings for track color, width, etc. (#418)
* changing the layering of two-city boom cities

The two-city boom city is assembled in the wrong order, putting the two
dashed-line circles atop the white oval.  This puts them behind it.

* add "match" color and globals for track color

This adds the ability to set track colors, width, etc. with globals in the
"info" section. Track can be made into a dashed line if the color and
borderColor are set the same as the hex color. However, it's inconvenient
to use them together since the hex color varies across the map and tiles.
This also adds a new color for track of "match" which means to match the
hex color.