chore(app): turn sandbox mode on
This means that we have better security in electron. The main and
preload scripts now render to common JS files and the windows created by
electron are now all sandboxed. The main process is still written as ESM
but the renderer is not since it's one require line isn't a real require
line anyway so there really is no difference.